multiple changes and improvement

+++ ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: pgeeser <pgeeser@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 22:50:07 by pgeeser           #+#    #+#             */
-/*   Updated: 2022/04/07 17:16:52 by pgeeser          ###   ########.fr       */
+/*   Updated: 2022/04/08 00:53:06 by pgeeser          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
 	if (!n)
 		return (0);
